fix: round-5 hunt findings — leaks/security/ABI/configs/dead-code/Go/Rust/warnings bulk
18 files, 19 distinct fixes across 11 of the 12 round-5 hunt cells
(race-conditions cell handled separately):

memory-leaks-strict (cell 2):
- core/src/libvmaf.c: set *vmaf=NULL after free(v) in free_v: path to
  prevent dangling pointer after vmaf_init sub-init failure
- core/src/thread_pool.c: set *pool=NULL after free(p) on workers-malloc
  failure; change if→while for pthread_cond_wait to guard spurious wakeups
- core/src/feature/feature_extractor.c: set *pool=NULL after free(p) at
  free_p: label; add comment documenting intentional mutex-init ordering

tiny-ai-model-runtime-audit (cell 9):
- core/src/pdjson.c: fix UBSan SIGABRT — change stack_top=-1 to
  (size_t)-1 at both init sites, matching the (size_t)-1 sentinel used
  at the two comparison sites (lines 720/746)

cross-extractor-numeric-comparison (cell 5):
- core/src/feature/x86/motion_avx512.c: fix signed integer overflow in
  three 16-bit scalar-tail paths (y_conv_8, y_conv_16, x_conv_16) — cast
  uint16_t filter[k] and src pixels to uint32_t before multiply so
  products stay in unsigned arithmetic (UBSan SIGABRT fix)

hidden-config-options (cell 6):
- core/meson_options.txt: fix inverted enable_nvcc description (was "Use
  clang…"; corrected to "Use nvcc…; when false, clang is used")
- docs/development/build-flags.md: remove false claim that enable_vulkan
  remains as a failing stub; the option() was fully deleted per ADR-0726
- core/src/meson.build: remove dead -DOC_NEW_STYLE_INCLUDES legacy define
  (no C consumer in tree since upstream libogg era)
- core/src/feature/adm.c: collapse ADM_OPT_SINGLE_PRECISION #ifdef branch
  (symbol never defined anywhere; 1e-2 threshold was permanently dead code)

unused-code-dead-code (cell 7):
- core/src/picture.h: remove dead VMAF_PICTURE_BUFFER_TYPE_VULKAN_DEVICE
  enum value (Vulkan removed per ADR-0726; no reference sites remain)

meson-build-warnings-grind (cell 12):
- core/src/opt.h: char*→const char* for default_val.s to prevent
  write-to-string-literal UB (icpx -Wwritable-strings)
- core/src/opt.c + opt.cpp: propagate const through set_option_string
  signature; opt.cpp uses const_cast per ADR-0721 ABI-stability comment
- core/src/feature/sycl/sycl_compat.h: update VMAF_SYCL_REQD_SG_SIZE macro
  from deprecated [[intel::reqd_sub_group_size(N)]] to SYCL 2020 standard
  [[sycl::reqd_sub_group_size(N)]] (oneAPI 2026.0 deprecation warning)

go-pkg-error-handling-audit (cell 10):
- pkg/libvmaf/libvmaf.go: check tmpOut.Close() error; use logged deferred
  remove instead of bare defer os.Remove
- cmd/vmafx-controller/queue/queue.go: surface json.Marshal error in
  ReportResult instead of silently falling back to "{}"; fix ListAll
  defer rows.Close() to log close errors (matching reload() pattern)
- cmd/vmafx-mcp/impl.go: intArg/floatArg return def (not 0) on
  json.Number conversion failure

rust-bindings-safety (cell 11):
- bindings/rust/vmafx-sys/tests/integration_test.rs: add debug_assert for
  non-negative strides before as-usize cast; add allow(cast_sign_loss)
  with justification comment

Build verified: ninja -C /tmp/vmaf-round5-build (CPU-only) 894/894 OK,
meson test --suite=fast 87/87 OK (including test_motion_avx512_parity
which previously aborted under UBSan). go build clean. cargo check clean.

+++ core/src/feature/x86/motion_avx512.c
@@ -465,11 +465,14 @@ void y_convolution_8_avx512(const void *src_void, uint16_t *dst, unsigned width,
                                 _mm512_cvtepi32_epi16(acc_hi));
         }
 
-        /* Scalar tail */
+        /* Scalar tail — cast to uint32_t before multiply to avoid signed int overflow
+         * (uint16_t * uint16_t promotes to signed int in C; at UINT16_MAX the product
+         * can exceed INT32_MAX, triggering UBSan). */
         for (; j < width; j++) {
-            uint32_t accum = (uint32_t)filter[0] * s0[j] + (uint32_t)filter[1] * s1[j] +
-                             (uint32_t)filter[2] * s2[j] + (uint32_t)filter[3] * s3[j] +
-                             (uint32_t)filter[4] * s4[j];
+            uint32_t accum =
+                (uint32_t)filter[0] * (uint32_t)s0[j] + (uint32_t)filter[1] * (uint32_t)s1[j] +
+                (uint32_t)filter[2] * (uint32_t)s2[j] + (uint32_t)filter[3] * (uint32_t)s3[j] +
+                (uint32_t)filter[4] * (uint32_t)s4[j];
             dst[i * dst_stride + j] = (uint16_t)((accum + add_before_shift) >> shift_var);
         }
     }
@@ -484,7 +487,7 @@ void y_convolution_8_avx512(const void *src_void, uint16_t *dst, unsigned width,
                     i_tap = -i_tap;
                 else if (i_tap >= (int)height)
                     i_tap = (int)height - (i_tap - (int)height + 2);
-                accum += filter[k] * src[(ptrdiff_t)i_tap * src_stride + j];
+                accum += (uint32_t)filter[k] * (uint32_t)src[(ptrdiff_t)i_tap * src_stride + j];
             }
             dst[i * dst_stride + j] = (uint16_t)((accum + add_before_shift) >> shift_var);
         }
@@ -547,11 +550,12 @@ void y_convolution_16_avx512(const void *src_void, uint16_t *dst, unsigned width
             _mm256_storeu_si256((__m256i *)(dst + i * dst_stride + j), _mm512_cvtepi32_epi16(acc));
         }
 
-        /* Scalar tail */
+        /* Scalar tail — cast to uint32_t before multiply; see y_convolution_8 comment. */
         for (; j < width; j++) {
-            uint32_t accum = (uint32_t)filter[0] * s0[j] + (uint32_t)filter[1] * s1[j] +
-                             (uint32_t)filter[2] * s2[j] + (uint32_t)filter[3] * s3[j] +
-                             (uint32_t)filter[4] * s4[j];
+            uint32_t accum =
+                (uint32_t)filter[0] * (uint32_t)s0[j] + (uint32_t)filter[1] * (uint32_t)s1[j] +
+                (uint32_t)filter[2] * (uint32_t)s2[j] + (uint32_t)filter[3] * (uint32_t)s3[j] +
+                (uint32_t)filter[4] * (uint32_t)s4[j];
             dst[i * dst_stride + j] = (uint16_t)((accum + add_before_shift) >> shift_var);
         }
     }
@@ -619,10 +623,13 @@ void x_convolution_16_avx512(const uint16_t *src, uint16_t *dst, unsigned width,
         }
 
         /* Scalar tail for remaining interior columns */
+        /* Scalar tail — cast to uint32_t before multiply; see y_convolution_8 comment. */
         for (; j < right_edge; j++) {
-            uint32_t accum = (uint32_t)filter[0] * src_row[0] + (uint32_t)filter[1] * src_row[1] +
-                             (uint32_t)filter[2] * src_row[2] + (uint32_t)filter[3] * src_row[3] +
-                             (uint32_t)filter[4] * src_row[4];
+            uint32_t accum = (uint32_t)filter[0] * (uint32_t)src_row[0] +
+                             (uint32_t)filter[1] * (uint32_t)src_row[1] +
+                             (uint32_t)filter[2] * (uint32_t)src_row[2] +
+                             (uint32_t)filter[3] * (uint32_t)src_row[3] +
+                             (uint32_t)filter[4] * (uint32_t)src_row[4];
             dst[i * dst_stride + j] = (uint16_t)((accum + shift_add_round) >> 16);
             src_row++;
         }
